Cleaned up generate_mask_comparison_figure.py.

+++ src/utils_image.py
@@ -135,8 +135,7 @@ def rgb_mask_to_one_hot(mask):
     mask_sums = mask.sum(axis=2)  # Sum of RGB values for each pixel; each of our colors happens to have a unique sum
     # The slice [:, :, np.newaxis] adds a third dimension, going from 480x480 to 480x480x1. This makes it possible to
     # do a broadcast comparison to COLOR_SUMS, giving 4 boolean values at each pixel, making the result 480x480x4.
-    return (mask_sums[:, :, np.newaxis] == rgb_mask_to_one_hot.COLOR_SUMS).astype(int)
-    # TODO Is int the type we want? uint8 will take less space, but maybe we want whatever comes out of the network
+    return (mask_sums[:, :, np.newaxis] == rgb_mask_to_one_hot.COLOR_SUMS).astype('uint8')
 
 
 # Sum of the RGB values for each color
